Provide an example of using a generic hook to theme a component
